Simplifying register functions
