+++ pumpkin/src/block/blocks/jukebox.rs
@@ -1,12 +1,10 @@
+use crate::block::pumpkin_block::{BlockMetadata, PumpkinBlock};
 use crate::entity::player::Player;
 use crate::server::Server;
-use crate::world::block::block_manager::{BlockManager, BlockUpdateConsumer, InteractiveBlock};
-use crate::world::block::pumpkin_block::{BlockMetadata, PumpkinBlock};
 use async_trait::async_trait;
 use pumpkin_core::math::position::WorldPosition;
 use pumpkin_registry::SYNCED_REGISTRIES;
 use pumpkin_world::item::item_registry::Item;
-use std::sync::Arc;
 
 pub struct JukeboxBlock;
 
@@ -15,16 +13,8 @@ impl BlockMetadata for JukeboxBlock {
     const ID: &'static str = "jukebox";
 }
 
-impl PumpkinBlock for JukeboxBlock {
-    fn register(self, block_manager: &mut BlockManager) {
-        let block = Arc::new(self);
-        block_manager.register_block_interactable(block.clone());
-        block_manager.register_with_block_update_consumer(block);
-    }
-}
-
 #[async_trait]
-impl InteractiveBlock for JukeboxBlock {
+impl PumpkinBlock for JukeboxBlock {
     async fn on_use<'a>(&self, player: &Player, location: WorldPosition, _server: &Server) {
         // For now just stop the music at this position
         let world = &player.living_entity.entity.world;
@@ -56,12 +46,5 @@ impl InteractiveBlock for JukeboxBlock {
         //TODO: Update block state and block nbt
 
         world.play_record(jukebox_song as i32, location).await;
-    }
-}
-
-#[async_trait]
-impl BlockUpdateConsumer for JukeboxBlock {
-    async fn on_placed<'a>(&self, _player: &Player, _location: WorldPosition, _server: &Server) {
-        todo!()
     }
 }
